Make precision a global, and move to its own file
